requirements.txt
# Base requirements
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
Werkzeug==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
APScheduler==3.10.1
PyJWT==2.6.0

# File handling
Pillow==9.5.0
python-magic==0.4.27

# Machine Learning
tensorflow==2.14.0
numpy==1.23.5
pandas==2.0.1
scikit-learn==1.2.2

# Testing
pytest==7.3.1
pytest-flask==1.2.0

FuzzyTM==1.0.0